Initial Plugin update for processing document tags

+++ crates/entities/src/models/indexed_document.rs
@@ -160,20 +160,45 @@ pub async fn insert_tags_for_docs<C: ConnectionTrait>(
         return Ok(());
     }
 
-    // Remove dupes before adding
-    let tags: HashSet<i64> = HashSet::from_iter(tags.iter().cloned());
     let doc_ids: Vec<i64> = docs.iter().map(|m| m.id).collect();
 
-    // Remove tags that are not in the tag set
-    let _ = document_tag::Entity::delete_many()
-        .filter(document_tag::Column::IndexedDocumentId.is_in(doc_ids.clone()))
-        .filter(document_tag::Column::TagId.is_not_in(tags.clone()))
-        .exec(db)
-        .await;
+    insert_tags_for_docs_by_id(db, &doc_ids, tags, true).await
+}
+
+/// Creates connections between a set of tags and a set of documents
+/// The document its provided are the document db id field and the
+/// tag ids are the tags database id field.
+///
+/// The remove unused option is utilized to remove any links to tags
+/// that are not in the list. This can be used when replacing the
+/// current set of tags with the passed in set
+pub async fn insert_tags_for_docs_by_id<C: ConnectionTrait>(
+    db: &C,
+    doc_ids: &[i64],
+    tags: &[i64],
+    remove_unused: bool,
+) -> Result<(), DbErr> {
+    // Nothing to do if we have no docs or tags
+    if doc_ids.is_empty() || tags.is_empty() {
+        return Ok(());
+    }
+
+    // Remove dupes before adding
+    let tags: HashSet<i64> = HashSet::from_iter(tags.iter().cloned());
+    let doc_ids: Vec<i64> = doc_ids.to_vec();
+
+    if remove_unused {
+        // Remove tags that are not in the tag set
+        let _ = document_tag::Entity::delete_many()
+            .filter(document_tag::Column::IndexedDocumentId.is_in(doc_ids.clone()))
+            .filter(document_tag::Column::TagId.is_not_in(tags.clone()))
+            .exec(db)
+            .await;
+    }
 
     // Grab existing tags
     let existing_tags = document_tag::Entity::find()
-        .filter(document_tag::Column::IndexedDocumentId.is_in(doc_ids))
+        .filter(document_tag::Column::IndexedDocumentId.is_in(doc_ids.clone()))
         .all(db)
         .await
         .unwrap_or_default()
@@ -184,11 +209,11 @@ pub async fn insert_tags_for_docs<C: ConnectionTrait>(
     // Only add tags that have not been added before.
     let tags = tags.sub(&existing_tags);
     // create connections for each tag
-    let doc_tags = docs
+    let doc_tags = doc_ids
         .iter()
-        .flat_map(|model| {
+        .flat_map(|id| {
             tags.iter().map(|t| document_tag::ActiveModel {
-                indexed_document_id: Set(model.id),
+                indexed_document_id: Set(*id),
                 tag_id: Set(*t),
                 created_at: Set(chrono::Utc::now()),
                 updated_at: Set(chrono::Utc::now()),
@@ -220,6 +245,23 @@ pub async fn insert_tags_for_docs<C: ConnectionTrait>(
             return Err(err);
         }
     }
+
+    Ok(())
+}
+
+/// Removes the specified tags from the specified documents. The ids for the
+/// tags and documents are the database id fields
+pub async fn remove_tags_for_docs_by_id<C: ConnectionTrait>(
+    db: &C,
+    doc_ids: &[i64],
+    tags: &[i64],
+) -> Result<(), DbErr> {
+    // Remove specified tags
+    let _ = document_tag::Entity::delete_many()
+        .filter(document_tag::Column::IndexedDocumentId.is_in(doc_ids.to_vec()))
+        .filter(document_tag::Column::TagId.is_in(tags.to_vec()))
+        .exec(db)
+        .await;
 
     Ok(())
 }
@@ -327,6 +369,60 @@ pub async fn find_by_lens(
         vec![name.into()],
     ))
     .all(&db)
+    .await
+}
+
+/// Helper method used to access the documents database id field from the
+/// string document id
+pub async fn find_by_doc_ids(
+    db: &DatabaseConnection,
+    ids: &[String],
+) -> Result<Vec<IndexedDocumentId>, sea_orm::DbErr> {
+    let doc_ids = ids
+        .iter()
+        .map(|str| format!("\"{str}\""))
+        .collect::<Vec<String>>()
+        .join(",");
+
+    IndexedDocumentId::find_by_statement(Statement::from_string(
+        DatabaseBackend::Sqlite,
+        format!(
+            r#"
+        SELECT
+            id,
+            doc_id
+        FROM indexed_document
+        WHERE doc_id in ({})"#,
+            doc_ids
+        ),
+    ))
+    .all(db)
+    .await
+}
+
+/// Represents the tag id that is associated with a document
+#[derive(Debug, FromQueryResult)]
+pub struct IndexedDocumentTagId {
+    pub id: i64,
+}
+
+/// Helper method used to access the database ids of the tags associated with the
+/// specified document. The passed in document id is the string document id field.
+pub async fn get_tag_ids_by_doc_id(
+    db: &DatabaseConnection,
+    id: &str,
+) -> Result<Vec<IndexedDocumentTagId>, sea_orm::DbErr> {
+    IndexedDocumentTagId::find_by_statement(Statement::from_sql_and_values(
+        DatabaseBackend::Sqlite,
+        r#"
+        SELECT
+            document_tag.tag_id as id
+        FROM document_tag as document_tag
+        LEFT JOIN indexed_document as indexed_doc on document_tag.indexed_document_id = indexed_doc.id
+        WHERE indexed_doc.doc_id = $1"#,
+        vec![id.into()],
+    ))
+    .all(db)
     .await
 }
